feat(cite): include original imprint in citation output
Footer feedback (Derek, 2026-06-04, Drebbel Tractatus duo): 'put the
bibliographic info in the cite info'. Citations previously credited only
Source Library as publisher and dropped the original imprint entirely.

- New src/lib/citations.ts: shared generateCitations + formatImprint.
  The global and tenant quote routes had byte-identical private copies
  (the #2420 drift pattern); both now import the shared module.
- All citation formats (APA, footnote, bibliography, Chicago, MLA,
  BibTeX) now carry 'Place: Publisher, Year' when the book records an
  imprint; BibTeX uses publisher/address fields with Source Library as
  translator.
- CiteButton already received publisher/placePublished props from both
  call sites but never used them — now it does.
